feat #265 : lock 개선
